+++ cmd/smpe_lint/main.go
@@ -15,7 +15,7 @@ import (
 )
 
 var (
-	version = "v0.8.6"
+	version = "v0.8.7"
 	commit  = "unknown"
 )